Use Win32 API calls to work around Activate issues
The `Activate` call will only bring the window to the foreground under
specific circumstances. In the case that another application is in the
foreground, the `Activate` call will just make the task bar icon flash.
Since the intent is to force the window to the foreground, use
SetForegroundWindow to focus the window and ShowWindow to ensure it's
visible.